refactor: extract test modules to bring state.rs and resolution_tests.rs under 2000 lines
Both files exceeded the 2000-line hard cap from CLAUDE.md.

state.rs (2028 → 1996 lines): Extract the ParserState overflow/reset unit
tests into state_tests.rs, referenced via #[cfg(test)] #[path] to preserve
the existing `use super::ParserState` scope without a mod.rs change.

resolution_tests.rs (2039 → 1914 lines): Extract the
jsdoc_import_type_specifier_collection_tests module into
resolution_tests_jsdoc.rs, using the same #[path] convention already
established by resolution_tests_path_mappings.rs and
resolution_tests_package_exports.rs.

No logic change; method bodies and test assertions are byte-for-byte identical.
